+++ packages/bot-kit-pro/src/bot.ts
