+++ packages/bot-kit-pro/src/config.ts
@@ -1,4 +1,4 @@
-import type { BotHandler } from "./foo.js"
+import type { BotHandler } from "./bot.js"
 import { getEnvOrDefault } from "./utils.js"
 
 const DEFAULT_XMTP_ENV = "dev"
