+++ packages/bot-kit-pro/src/index.ts
@@ -2,4 +2,4 @@ export { default as run } from "./run.js"
 export { default as HandlerContext } from "./context.js"
 export { BotConfig, newBotConfig } from "./config.js"
 export { Json } from "./types.js"
-export { default as Bot } from "./foo.js"
+export { default as Bot } from "./bot.js"
